+++ mlos_bench/mlos_bench/tests/storage/trial_data_test.py
@@ -10,11 +10,6 @@ import pytest
 
 from mlos_bench.environments.status import Status
 from mlos_bench.storage.base_experiment_data import ExperimentData
-from mlos_bench.tests.storage import (
-    CONFIG_COUNT,
-    CONFIG_TRIAL_REPEAT_COUNT,
-    TRIAL_RUNNER_COUNT,
-)
 
 
 def test_exp_trial_data(exp_data: ExperimentData) -> None:
@@ -34,18 +29,3 @@ def test_exp_trial_data(exp_data: ExperimentData) -> None:
     assert isinstance(trial.ts_start, datetime)
     assert isinstance(trial.ts_end, datetime)
     # Note: tests for telemetry are in test_update_telemetry()
-
-
-def test_rr_scheduling(exp_data: ExperimentData) -> None:
-    """Checks that the scheduler produced basic round-robin scheduling of Trials across
-    Runners.
-    """
-    for trial_id in range(1, CONFIG_COUNT * CONFIG_TRIAL_REPEAT_COUNT):
-        expected_config_id = trial_id // CONFIG_TRIAL_REPEAT_COUNT + 1
-        expected_repeat_num = (trial_id - 1) % CONFIG_TRIAL_REPEAT_COUNT + 1
-        expected_runner_id = (trial_id - 1) % TRIAL_RUNNER_COUNT + 1
-        trial = exp_data.trials[trial_id]
-        assert trial.trial_id == trial_id
-        assert trial.tunable_config_id == expected_config_id
-        assert trial.metadata_dict["repeat_i"] == expected_repeat_num
-        assert trial.trial_runner_id == expected_runner_id
