+++ mlos_bench/mlos_bench/tests/storage/trial_schedule_test.py
@@ -9,7 +9,13 @@ from datetime import datetime, timedelta
 from pytz import UTC
 
 from mlos_bench.environments.status import Status
+from mlos_bench.storage.base_experiment_data import ExperimentData
 from mlos_bench.storage.base_storage import Storage
+from mlos_bench.tests.storage import (
+    CONFIG_COUNT,
+    CONFIG_TRIAL_REPEAT_COUNT,
+    TRIAL_RUNNER_COUNT,
+)
 from mlos_bench.tunables.tunable_groups import TunableGroups
 
 
@@ -116,3 +122,18 @@ def test_schedule_trial(exp_storage: Storage.Experiment, tunable_groups: Tunable
     assert trial_ids == [trial_1h.trial_id]
     assert len(trial_configs) == len(trial_scores) == 1
     assert trial_status == [Status.SUCCEEDED]
+
+
+def test_rr_scheduling(exp_data: ExperimentData) -> None:
+    """Checks that the scheduler produced basic round-robin scheduling of Trials across
+    Runners.
+    """
+    for trial_id in range(1, CONFIG_COUNT * CONFIG_TRIAL_REPEAT_COUNT):
+        expected_config_id = trial_id // CONFIG_TRIAL_REPEAT_COUNT + 1
+        expected_repeat_num = (trial_id - 1) % CONFIG_TRIAL_REPEAT_COUNT + 1
+        expected_runner_id = (trial_id - 1) % TRIAL_RUNNER_COUNT + 1
+        trial = exp_data.trials[trial_id]
+        assert trial.trial_id == trial_id
+        assert trial.tunable_config_id == expected_config_id
+        assert trial.metadata_dict["repeat_i"] == expected_repeat_num
+        assert trial.trial_runner_id == expected_runner_id
